requirements.txt
openai == 0.28
fpdf
python-dotenv